Update pipeline and agents w.r.t. InFormALin 0.9.0
