+++ pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <groupId>io.github.informalin.framework</groupId>
     <artifactId>parent</artifactId>
-    <version>0.8.1</version>
+    <version>0.9.0</version>
   </parent>
 
   <groupId>io.github.ardoco.core</groupId>
@@ -14,12 +14,12 @@
   <packaging>pom</packaging>
   <name>ArDoCo (Core) - The Consistency Analyzer</name>
   <description>The goal of this project is to connect architecture documentation and models while identifying missing
-        or deviating elements (inconsistencies). An element can be any representable item of the model, like a component
-        or a relation. To do so, we first create trace links and then make use of them and other information to identify
-        inconsistencies. ArDoCo is actively developed by researchers of the Modelling for Continuous Software Engineering
-        (MCSE) group of KASTEL - Institute of Information Security and Dependability at the KIT. This work was supported by
-        funding from the topic Engineering Secure Systems of the Helmholtz Association (HGF) and by KASTEL Security Research
-        Labs (46.23.01).</description>
+    or deviating elements (inconsistencies). An element can be any representable item of the model, like a component
+    or a relation. To do so, we first create trace links and then make use of them and other information to identify
+    inconsistencies. ArDoCo is actively developed by researchers of the Modelling for Continuous Software Engineering
+    (MCSE) group of KASTEL - Institute of Information Security and Dependability at the KIT. This work was supported by
+    funding from the topic Engineering Secure Systems of the Helmholtz Association (HGF) and by KASTEL Security Research
+    Labs (46.23.01).</description>
   <licenses>
     <license>
       <name>MIT License</name>
